kafka-uber-demo/client/requirements.txt
kafka-python==2.0.2
openai==1.31.0
tqdm
python-dotenv

